test_requirements.txt
# This file was generated by running:
#   sudo docker run --rm -it --network=host python:3.8-slim /bin/bash
#   pip install dnspython idna # from setup.cfg
#   pip install pytest pytest-asyncio pytest-cov coverage flake8 mypy
#   pip freeze
# (Some packages' latest versions may not be compatible with
# the earliest Python version we support, and some exception
# messages may depend on package versions, so we pin versions
# for reproducible testing.)
coverage==7.4.4
dnspython==2.6.1
exceptiongroup==1.2.0
flake8==7.0.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
tomli==2.0.1
typing_extensions==4.11.0
